+++ CHANGELOG.md
@@ -1,4 +1,16 @@
 
+# 0.5.2
+* Support numeric enumerator codes in JSONPB,
+  as required by the protobuf standard.
+* Handle unrecognized enumerator codes within packed repeated
+  enumeration fields in the same way as in other enumeration fields.
+* Avoid compilation errors in code generated for messages having BytesValue
+  fields.  The errors would trigger only when the "swagger-wrapper-format"
+  Cabal flag of this package was False.
+* Code generated from a .proto file that imports the google.protobuf package
+  no longer depends upon any Haskell module generated from "wrappers.proto".
+  Instead the proto3-suite library provides the necessary functionality.
+
 # 0.5.1
 * Support newer versions of proto3-wire, bytestring, and turtle
 * Increase minimum version of base for canonicalize-proto-file from 4.8 to 4.11